fixed Uptime rounding error and getCpuUtilizatioon

+++ src/system.cpp
@@ -44,4 +44,4 @@ int System::RunningProcesses() { return LinuxParser::RunningProcesses(); }
 int System::TotalProcesses() { return LinuxParser::TotalProcesses(); }
 
 // TODO: Return the number of seconds since the system started running
-long int System::UpTime() { return LinuxParser::UpTime(); }
+long int System::UpTime() { return static_cast<long int>(LinuxParser::UpTime()); }
